[medium/large] Split business logic out of App.jsx
One important thing you can try doing is keeping significant
"business logic" separate from your presentational/display logic.
This commit doesn't do much except extract a function from inside
App.jsx to outside, but what this lets you do is

a) Physically separate the code so it's not as complex (App.jsx is
   starting to get a little long with all the big functions inside
   of it)
b) Mentally separate the responsibilities. The App component holds
   all the state, but this way you've separated the idea of how a
   product is added to a cart out of the code that handles passing
   the data around to child components to display. You can think
   of it like this: the idea of adding a product to a cart has
   nothing to do with React, but is related to how your business
   (in this case, the storefront) works, so you should separate
   the business code from the React code.
c) It allows you to write tests for it later. Suppose your code
   gets even more complicated - if your core logic is separated,
   you can easily call the function with certain inputs and look
   for the correct expected outputs without the behaviour being
   buried inside of a React component.
